Casting fixes
Former-commit-id: ba8adb0d7b2142cfb3779a99172b4db2c2c49961
Former-commit-id: 8270c5990b5025877ed3bea4487a37f0c734e4f8

+++ Sources/FlounderEngine/Models/Shapes/Rectangle.hpp
@@ -26,8 +26,8 @@ namespace Flounder
 		static Rectangle *Resource(const std::string &filename)
 		{
 			auto split = FormatString::Split(filename, "_");
-			float width = atof(split.at(1).c_str());
-			float height = atof(split.at(2).c_str());
+			float width = static_cast<float>(atof(split.at(1).c_str()));
+			float height = static_cast<float>(atof(split.at(2).c_str()));
 			return Resource(width, height);
 		}
